LPS-179710 Rename fields in MDf Claim
