+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/routes/MDFClaimForm/MDFClaimForm.tsx
@@ -15,9 +15,6 @@ import PRMFormik from '../../common/components/PRMFormik';
 import {PRMPageRoute} from '../../common/enums/prmPageRoute';
 import useLiferayNavigate from '../../common/hooks/useLiferayNavigate';
 import MDFClaimDTO from '../../common/interfaces/dto/mdfClaimDTO';
-import MDFRequestActivityDTO from '../../common/interfaces/dto/mdfRequestActivityDTO';
-import LiferayPicklist from '../../common/interfaces/liferayPicklist';
-import MDFClaim from '../../common/interfaces/mdfClaim';
 import {Liferay} from '../../common/services/liferay';
 import useGetDocumentFolder from '../../common/services/liferay/headless-delivery/useGetDocumentFolders';
 import useGetMDFClaimById from '../../common/services/liferay/object/mdf-claim/useGetMDFClaimById';
@@ -27,98 +24,8 @@ import {getMDFClaimFromDTO} from '../../common/utils/dto/mdf-claim/getMDFClaimFr
 import MDFClaimPage from './components/MDFClaimPage';
 import claimSchema from './components/MDFClaimPage/schema/yup';
 import useGetMDFRequestIdByHash from './hooks/useGetMDFRequestIdByHash';
+import getInitialFormValues from './utils/getInitialFormValues';
 import submitForm from './utils/submitForm';
-
-const getInitialFormValues = (
-	mdfRequestId: number,
-	currency: LiferayPicklist,
-	activitiesDTO?: MDFRequestActivityDTO[],
-	totalMDFRequestAmount?: number,
-	mdfClaim?: MDFClaim
-): MDFClaim => ({
-	...mdfClaim,
-	activities: activitiesDTO?.map((activity) => {
-		const mdfClaimActivity = mdfClaim?.activities?.find(
-			(claimActivity) =>
-				claimActivity.r_actToMDFClmActs_c_activityId === activity.id
-		);
-
-		if (mdfClaimActivity) {
-			return {
-				...mdfClaimActivity,
-				activityStatus: activity.activityStatus,
-				budgets: activity?.actToBgts?.map((budget) => {
-					const mdfClaimBudget = mdfClaimActivity.budgets?.find(
-						(claimBudget) =>
-							claimBudget.r_bgtToMDFClmBgts_c_budgetId ===
-							budget.id
-					);
-
-					if (mdfClaimBudget) {
-						return {
-							...mdfClaimBudget,
-							r_bgtToMDFClmBgts_c_budgetId: budget.id,
-							requestAmount: budget.cost,
-						};
-					}
-
-					return {
-						expenseName: budget.expense.name,
-						invoiceAmount: budget.cost,
-						r_bgtToMDFClmBgts_c_budgetId: budget.id,
-						requestAmount: budget.cost,
-						selected: false,
-					};
-				}),
-				claimed: activity.actToMDFClmActs
-					?.map((mdfClaimActivity) => {
-						return (
-							mdfClaimActivity?.r_mdfClmToMDFClmActs_c_mdfClaim
-								?.mdfClaimStatus.key !== 'draft'
-						);
-					})
-					.includes(true),
-				name: activity.name,
-				r_actToMDFClmActs_c_activityId: activity.id,
-			};
-		}
-
-		return {
-			activityStatus: activity.activityStatus,
-			budgets: activity?.actToBgts?.map((budget) => {
-				return {
-					expenseName: budget.expense.name,
-					invoiceAmount: budget.cost,
-					r_bgtToMDFClmBgts_c_budgetId: budget.id,
-					requestAmount: budget.cost,
-					selected: false,
-				};
-			}),
-			claimed: activity.actToMDFClmActs
-				?.map((mdfClaimActivity) => {
-					return (
-						mdfClaimActivity?.r_mdfClmToMDFClmActs_c_mdfClaim
-							?.mdfClaimStatus.key !== 'draft'
-					);
-				})
-				.includes(true),
-			currency: activity.currency,
-			metrics: '',
-			name: activity.name,
-			r_actToMDFClmActs_c_activityId: activity.id,
-			selected: false,
-			totalCost: 0,
-		};
-	}),
-	currency: mdfClaim?.currency ? mdfClaim?.currency : currency,
-	mdfClaimStatus: mdfClaim?.mdfClaimStatus
-		? mdfClaim.mdfClaimStatus
-		: Status.PENDING,
-	r_mdfReqToMDFClms_c_mdfRequestId: mdfRequestId,
-	totalMDFRequestedAmount: mdfClaim?.totalMDFRequestedAmount
-		? mdfClaim.totalMDFRequestedAmount
-		: totalMDFRequestAmount,
-});
 
 const SECOND_POSITION_AFTER_HASH = 1;
 const FOURTH_POSITION_AFTER_HASH = 3;
