+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/routes/MDFRequestForm/utils/createMDFRequestProxyAPI.ts
@@ -40,6 +40,7 @@ export default async function createMDFRequestProxyAPI(mdfRequest: MDFRequest) {
 			dtoMDFRequestResponse = await updateMDFRequest(
 				ResourceName.MDF_REQUEST_DXP,
 				mdfRequest,
+				mdfRequest.id,
 				dtoMDFRequestSFResponse.externalReferenceCode
 			);
 		}
