+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/src/main/resources/site-initializer/fragments/group/src/mdf-claim-detail-collection/fragments/claim-detail-activity-list/index.html
@@ -10,11 +10,9 @@
 			<div>
 				<div class="d-flex">
 					<div class="w-100">
-						<h4 class="text-neutral-10">${activity.name}</h4>
+						<h5 class="text-neutral-10">${activity.name}</h5>
 
 						<div class="d-flex justify-content-between">
-							<h6 class="text-neutral-6">MDF Claim:</h6>
-
 							<h5 class="justify-content-end text-neutral-10">${get_currency(activity.totalCost?number, (activity.currency.key)!"USD")}</h5>
 						</div>
 					</div>
@@ -24,6 +22,16 @@
 
 		<div class="collapse panel-collapse" id="mdfClaimActivitiesPanel-${activity.id}" role="tabpanel">
 			<div class="panel-body">
+				<a class="align-items-end d-flex mx-2 my-3 text-decoration-none" href="${activity.listOfQualifiedLeads.link.href}">
+					<span class="mr-2 mt-3 text-neutral-5">
+						[@clay.icon class="mr-2 mt-3 text-neutral-5" symbol="file-xsl" /]
+					</span>
+
+					<div>
+						<h5 class="mb-0 text-neutral-9">${activity.listOfQualifiedLeads.name?keep_before("#")}</h5>
+					</div>
+				</a>
+
 				<div class="table-responsive">
 					<table class="bg-brand-primary-lighten-6 border-0 show-quick-actions-on-hover table table-autofit table-hover table-list table-striped">
 						<thead>
@@ -38,11 +46,20 @@
 						<tbody>
 							[#list activity.mdfClmActToMDFClmBgts as budget]
 								<tr>
-									<td class="border-0 w-50">
+									<td class="border-0">
 										<p class="text-neutral-10">${budget.expenseName}</p>
-									</td>
-									<td class="border-0 w-50">
 										<p class="text-neutral-10">${get_currency(budget.invoiceAmount?number, (activity.currency.key)!"USD")}</p>
+									</td>
+									<td class="align-baseline border-0">
+										<a class="align-items-end d-flex mx-2 my-3 text-decoration-none" href="${budget.invoice.link.href}">
+											<span class="mr-2 mt-3 text-neutral-5">
+												[@clay.icon class="mr-2 mt-3 text-neutral-5" symbol="document-table" /]
+											</span>
+
+											<div>
+												<h5 class="mb-0 text-neutral-9">${budget.invoice.name?keep_before("#")}</h5>
+											</div>
+										</a>
 									</td>
 								</tr>
 							[/#list]
