+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/src/main/resources/site-initializer/fragments/group/src/mdf-claim-detail-collection/fragments/claim-detail-document-list/index.html
@@ -8,28 +8,23 @@
 			</p>
 
 			<div class="card-text">
-				<a class="d-flex mt-3 text-decoration-none" href="${document.url}">
+				<a class="align-items-end d-flex mt-3 text-decoration-none" href="${document.link.href}">
 					<span class="mr-2 mt-3 text-neutral-5">
 						[@clay.icon class="mr-2 mt-3 text-neutral-5" symbol="document-pdf" /]
 					</span>
 
 					<div>
-						<h5 class="mb-0 text-neutral-9">${document.fileName}</h5>
-
-						<p class="mb-0 text-neutral-9">Size: ${(document.fileSize?number/1000)?ceiling}KB</p>
+						<h5 class="mb-0 text-neutral-9">${document.name?keep_before("#")}</h5>
 					</div>
 				</a>
 			</div>
 		</div>
 	</div>
 [/#macro]
-
 [#if mdf_claim_id?has_content]
-	[#assign documents = restClient.get("/c/mdfclaims/${mdf_claim_id}/mdfClmToMDFClmDocs").items /]
+	[#assign claim = restClient.get("/c/mdfclaims/${mdf_claim_id}") /]
 
-	[#list documents as document]
-		[@document_card document=document /]
-	[/#list]
+	[@document_card document=claim.reimbursementInvoice /]
 [#else]
 	<div class="alert alert-info">
 		<p>To use this fragment either place it on a Display Page Template to see which attributes are available from <em>request.getAttribute("INFO_ITEM")</em>.</p>
