+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/src/main/resources/site-initializer/object-definitions/mdf-claim.json
@@ -151,7 +151,8 @@
 							},
 							{
 								"key": "claimPaid",
-								"objectStateTransitions": []
+								"objectStateTransitions": [
+								]
 							},
 							{
 								"key": "inFinanceReview",
